Move ParseAST
Move ParseAST to the Macronify binary's directory since only that binary
uses this function now.
