+++ bin/Macronify/ParseAST.cpp
@@ -1,4 +1,4 @@
-#include <macroni/Common/ParseAST.hpp>
+#include "ParseAST.hpp"
 #include <pasta/Compile/Command.h>
 #include <pasta/Compile/Compiler.h>
 #include <pasta/Util/ArgumentVector.h>
